+++ backend/src/main/java/io/cooktail/backend/domain/cocktail/repository/CocktailImageRepository.java
@@ -1,6 +1,6 @@
-package io.cooktail.backend.cocktail.repository;
+package io.cooktail.backend.domain.cocktail.repository;
 
-import io.cooktail.backend.cocktail.domain.CocktailImage;
+import io.cooktail.backend.domain.cocktail.domain.CocktailImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
